feat: complete booking module implementation

+++ backend/rental-api/src/main/java/run/frank/rentalapi/entity/Booking.java
@@ -21,7 +21,7 @@ public class Booking {
 
     private Long tenantId;
 
-    private Integer propertyId;
+    private Long propertyId;
 
     private LocalDate startDate;
